+++ src/lib/tools/export.ts
@@ -8,7 +8,7 @@ import valueDiff from './valueDiff';
 import listItems from './summarizeItems';
 import summarize, { summarizeToChat } from './summarizeOffer';
 import profit from './profit';
-import itemStats from './item-stats';
+import itemStats from './itemStats';
 
 export {
     getHighValueItems,
