+++ src/lib/tools/itemStats.ts
